travis: Install mopidy==dev in virtualenv
The `mopidy` package from APT ensures that all dependencies are in place. The APT package installs the Python code into /usr/share/mopidy and are thus not available on PYTHONPATH. Installing Mopidy from PyPI is more convenient than messing with PYTHONPATH. Also, we want the development of extensions to follow Mopidy's development, not the last release in APT.

Fixes #1.